Typo fix, PEP8 fixes, move FrozenSourceReservoir stuff to its own file.
