+++ flamedisx/__init__.py
@@ -8,6 +8,7 @@ from .likelihood import *
 from .inference import *
 from .bounds import *
 from .mu_estimation import *
+from .frozen_reservoir import *
 
 # Original flamedisx models
 # Accessible under fd root package (for now), for backwards compatibility
